Remove SSO-app and django-oauth-toolkit
