+++ apps/online_oidc_provider/claims.py
@@ -1,7 +1,7 @@
 from django.utils.translation import gettext as _
 from oidc_provider.lib.claims import ScopeClaims
 
-from apps.sso.userinfo import Onlineweb4Userinfo
+from apps.api.userinfo import Onlineweb4Userinfo
 
 
 def userinfo(claims, user):
