--- apps/sso/templatetags/zip_lists.py
@@ -1,8 +1,0 @@
-from django import template
-
-register = template.Library()
-
-
-@register.filter
-def zip_lists(a, b):
-    return zip(a, b)
