+++ onlineweb4/settings/rest_framework.py
@@ -2,9 +2,6 @@
 REST_FRAMEWORK = {
     "DEFAULT_SCHEMA_CLASS": "apps.api.utils.PrefixRemovedAutoSchema",
     "DEFAULT_AUTHENTICATION_CLASSES": (
-        # Allows users to be logged in with open-id through django-oauth-toolkit
-        # Has to be listed before OidcOauth2Auth!!
-        "oauth2_provider.contrib.rest_framework.OAuth2Authentication",
         # Allows user to be logged in with open-id
         "apps.online_oidc_provider.authentication.OidcOauth2Auth",
         # Allows users to be logged in to browsable API
