--- templates/profiles/connected_apps.html
@@ -1,48 +1,0 @@
-<div class="row">
-    <div class="col-md-12">
-        <h4>Aktive applikasjonstilkoblinger</h4>
-    </div>
-</div>
-<div class="row">
-    <div class="col-md-12">
-        <p>Her finner du en oversikt over alle eksterne applikasjoner du har godkjent.
-        Standard varighet for en tilgangsnøkkel er vanligvis 60 minutter, men kan forlenges
-        dersom applikasjonen ønsker det. Du kan når som helst stenge apptilkoblingen.</p>
-    </div>
-    <div class="col-md-12">
-        {% if connected_apps %}
-        <table class="table table-striped-table-condensed">
-            <thead>
-            <tr>
-                <th>Applikasjon</th>
-                <th>Utløpstid</th>
-                <th>Tilganger</th>
-                <th class="text-right">Steng</th>
-            </tr>
-            </thead>
-            <tbody>
-            {% for token in connected_apps %}
-            <tr>
-                <td>{{ token.application.name }}</td>
-                <td>{{ token.expires|date:'d.m.y H:i:s' }}</td>
-                <td>
-                    {% for scope in token.scope.split %}
-                        {{ scope }}<br/>
-                    {% endfor %}
-                </td>
-                <td>
-                    <form method="POST" action="{% url 'profile_connected_apps' %}">
-                        {% csrf_token %}
-                        <input type="hidden" name="token_id" value="{{ token.id }}" />
-                        <button type="submit" class="btn btn-danger pull-right">Steng</button>
-                    </form>
-                </td>
-            </tr>
-            {% endfor %}
-            </tbody>
-        </table>
-        {% else %}
-        <p><i>Ingen aktive apptilkoblinger.</i></p>
-        {% endif %}
-    </div>
-</div>
