--- templates/sso/dashboard/app_details.html
@@ -1,107 +1,0 @@
-{% extends "dashboard_base.html" %}
-
-{% block title %}
-{{ app.name }} - Ekstern tilgang - Online Dashboard
-{% endblock title %}
-
-{% block page-header %}
-{{ app.name }}
-{% endblock %}
-
-{% block breadcrumbs %}
-    <li><a href="/dashboard/auth/">Autentisering</a></li>
-    <li><a href="{% url 'sso:index' %}">Ekstern tilgang</a></li>
-    <li>{{ app.name }}</li>
-{% endblock %}
-
-{% block content %}
-<div class="row">
-    <div class="col-md-12">
-        <a href="{% url 'sso:new_app' %}" class="btn btn-primary pull-left">
-            <i class="fa fa-plus"></i> Ny App-klient
-        </a>
-        <form method="POST">
-            {% csrf_token %}
-            <input type="hidden" name="action" value="delete">
-            <button type="submit" class="btn btn-danger pull-right">
-                <i class="fa fa-times"></i> Slett
-            </button>
-        </form>
-    </div>
-</div>
-<br />
-<div class="row">
-    <div class="col-md-12">
-        <div class="panel panel-default">
-            <div class="panel-heading">
-                <h3 class="panel-title">Applikasjonsoversikt</h3>
-            </div>
-            <div class="panel-body">
-                <div class="row">
-                    <div class="col-xs-12 col-sm-4 col-md-3 col-lg-2">
-                        <p><strong>Opprettet av:</strong></p>
-                    </div>
-                    <div class="col-xs-12 col-sm-8 col-md-9 col-lg-10">
-                        <p>{{ app.user.get_full_name }}</p>
-                    </div>
-                </div>
-                <div class="row">
-                    <div class="col-xs-12 col-sm-4 col-md-3 col-lg-2">
-                        <p><strong>Klient ID:</strong></p>
-                    </div>
-                    <div class="col-xs-12 col-sm-8 col-md-9 col-lg-10">
-                        <p style="word-wrap: break-word">{{ app.client_id }}</p>
-                    </div>
-                </div>
-                <div class="row">
-                    <div class="col-xs-12 col-sm-4 col-md-3 col-lg-2">
-                        <p><strong>Klient secret:</strong></p>
-                    </div>
-                    <div class="col-xs-12 col-sm-8 col-md-9 col-lg-10">
-                        <p style="word-wrap: break-word">{{ app.client_secret }}</p>
-                    </div>
-                </div>
-                <div class="row">
-                    <div class="col-xs-12 col-sm-4 col-md-3 col-lg-2">
-                        <p><strong>Klient type:</strong></p>
-                        <p><strong>Grant type:</strong></p>
-                        <p><strong>Redirect URIs:</strong></p>
-                    </div>
-                    <div class="col-xs-12 col-sm-8 col-md-9 col-lg-10">
-                        <p>{{ app.client_type }}</p>
-                        <p>{{ app.authorization_grant_type }}</p>
-                        <p>
-                            {% for uri in app.get_redirect_uris %}
-                                {{ uri }}<br />
-                            {% endfor %}
-                        </p>
-                    </div>
-                </div>
-            </div>
-        </div>
-        <div class="panel panel-default">
-            <div class="panel-heading">
-                <h3 class="panel-title">Tilganger</h3>
-            </div>
-            <div class="panel-body">
-                <table class="table table-striped table-condensed">
-                    <thead>
-                    <tr>
-                        <th>Beskrivelse</th>
-                        <th>Scope</th>
-                    </tr>
-                    </thead>
-                    <tbody>
-                    {% for scope, description in scopes %}
-                    <tr>
-                        <td>{{ description }}</td>
-                        <td>{{ scope }}</td>
-                    </tr>
-                    {% endfor %}
-                    </tbody>
-                </table>
-            </div>
-        </div>
-    </div>
-</div>
-{% endblock content %}
