--- templates/sso/dashboard/index.html
@@ -1,88 +1,0 @@
-{% extends "dashboard_base.html" %}
-
-{% block title %}
-Ekstern tilgang - Online Dashboard
-{% endblock title %}
-
-{% block page-header %}
-Ekstern Tilgang
-{% endblock %}
-
-{% block breadcrumbs %}
-    <li><a href="/dashboard/auth/">Autentisering</a></li>
-    <li>Ekstern tilgang</li>
-{% endblock %}
-
-{% block content %}
-<div class="row">
-    <div class="col-md-12">
-        <a href="{% url 'sso:new_app' %}" class="btn btn-primary">
-            <i class="fa fa-plus"></i> Ny applikasjon
-        </a>
-    </div>
-</div>
-<br />
-<div class="row">
-    <div class="col-md-12">
-        <div class="panel panel-default">
-            <div class="panel-heading">
-                <h3 class="panel-title">Applikasjonsoversikt</h3>
-            </div>
-            <div class="panel-body">
-                <p>Oversikt over eksterne applikasjoner</p>
-                <table class="table table-striped table-condensed">
-                    <thead>
-                    <tr>
-                        <th>Name</th>
-                        <th>Application ID</th>
-                        <th>Redirect URIs</th>
-                        <th>Grant type</th>
-                    </tr>
-                    </thead>
-                    <tbody>
-                    {% for app in apps %}
-                    <tr>
-                        <td>
-                            <a href="{% url 'sso:app_details' app.id %}">{{ app.name }}</a>
-                        </td>
-                        <td>{{ app.client_id }}</td>
-                        <td>{{ app.redirect_uris }}</td>
-                        <td>{{ app.authorization_grant_type }}</td>
-                    </tr>
-                    {% endfor %}
-                    </tbody>
-                </table>
-            </div>
-        </div>
-        <div class="panel panel-default">
-            <div class="panel-heading">
-                <h3 class="panel-title">Tilgjengelige scopes</h3>
-            </div>
-            <div class="panel-body">
-                <p>Følgende liste over tilgangsnivåer er tilgjengelig for bruk under registrering av nye applikasjoner.
-                Vær oppmerksom på at det ikke nødvendigvis er implementert noen endepunkter som faktisk gir ut disse
-                dataene. Django OAuth2 Toolkit behøver et forhåndsdefinert sett med scopes for å binde til tilgansngøklene,
-                og dermed kan de igjen brukes i decorators for å begrense tilgang til views.</p>
-                <p>For å legge til flere scopes må man legge til disse i <i>settings.py</i> under
-                <i>OAUTH2_PROVIDER</i>.</p>
-                <table class="table table-striped table-condensed">
-                    <thead>
-                        <tr>
-                            <th>Beskrivelse</th>
-                            <th>Scope</th>
-                        </tr>
-                    </thead>
-                    <tbody>
-                    {% for scope_value, scope_description in available_scopes %}
-                    <tr>
-                        <td>{{ scope_description }}</td>
-                        <td>{{ scope_value }}</td>
-                    </tr>
-                    {% endfor %}
-                    </tbody>
-                </table>
-            </div>
-        </div>
-    </div>
-</div>
-{% endblock content %}
