--- templates/sso/dashboard/new_app.html
@@ -1,53 +1,0 @@
-{% extends "dashboard_base.html" %}
-{% load addcss %}
-{% load crispy_forms_tags %}
-
-{% block title %}
-Ny App-klient - Ekstern tilgang - Online Dashboard
-{% endblock title %}
-
-{% block page-header %}
-Ny App-klient
-{% endblock %}
-
-{% block breadcrumbs %}
-    <li><a href="/dashboard/auth/">Autentisering</a></li>
-    <li><a href="{% url 'sso:index' %}">Ekstern tilgang</a></li>
-    <li>Ny App-klient</li>
-{% endblock %}
-
-{% block content %}
-<div class="row">
-    <div class="col-md-12">
-        <a class="btn btn-danger" href="{% url 'sso:index' %}">
-            <i class="fa fa-times"></i> Avbryt
-        </a>
-    </div>
-</div>
-<br />
-<div class="row">
-    <div class="col-md-12">
-        <div class="panel panel-default">
-            <div class="panel-heading">
-                <h3 class="panel-title">Ny App-klient</h3>
-            </div>
-            <div class="panel-body">
-                <p><strong>NB!</strong> Både godkjente Redirect-URI'er og tilganger (Scopes) må lagres space-separated.</p>
-                <form role="form" method="POST">
-                {% csrf_token %}
-                {% for field in form %}
-                <div class="form-group {% if field.errors %}has-error{% endif %}">
-                    <label for="{{ field.auto_id }}">{{ field.label }}</label>
-                    {{ field|addclass:"form-control" }}
-                    {% if field.errors %}
-                    <p>{{ field.errors }}</p>
-                    {% endif %}
-                </div>
-                {% endfor %}
-                <button type="submit" class="btn btn-success"><i class="fa fa-save"></i> Lagre</button>
-                </form>
-            </div>
-        </div>
-    </div>
-</div>
-{% endblock content %}
